+++ integration_tests/suite/base/test_iax_callnumberlimits.py
@@ -1,10 +1,11 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from hamcrest import assert_that, has_entries
 
 from ..helpers import scenarios as s
 from . import confd
+from ..helpers.config import TOKEN_SUB_TENANT
 
 
 def test_put_errors():
@@ -92,6 +93,14 @@ def test_edit_iax_callnumberlimits_without_items():
 
     response = confd.asterisk.iax.callnumberlimits.get()
     assert_that(response.item, has_entries(parameters))
+
+
+def test_restrict_only_master_tenant():
+    response = confd.asterisk.iax.callnumberlimits.get(token=TOKEN_SUB_TENANT)
+    response.assert_status(401)
+
+    response = confd.asterisk.iax.callnumberlimits.put(token=TOKEN_SUB_TENANT)
+    response.assert_status(401)
 
 
 def test_bus_event_when_edited():