+++ integration_tests/suite/base/test_pjsip_global.py
@@ -1,10 +1,11 @@
-# Copyright 2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2020-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from hamcrest import assert_that, has_entries
 
 from . import confd
 from ..helpers import scenarios as s
+from ..helpers.config import TOKEN_SUB_TENANT
 
 
 def test_put_errors():
@@ -48,6 +49,14 @@ def test_edit_with_no_option():
 
     response = confd.asterisk.pjsip.global_.get()
     assert_that(response.item, has_entries(parameters))
+
+
+def test_restrict_only_master_tenant():
+    response = confd.asterisk.pjsip.global_.get(token=TOKEN_SUB_TENANT)
+    response.assert_status(401)
+
+    response = confd.asterisk.pjsip.global_.put(token=TOKEN_SUB_TENANT)
+    response.assert_status(401)
 
 
 def test_bus_event_when_edited():