+++ integration_tests/suite/helpers/client.py
@@ -1,4 +1,4 @@
-# Copyright 2015-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import csv
@@ -118,28 +118,45 @@ class RestUrlClient(UrlFragment):
         params = self._merge_params(params, self.body)
         return self.client.head(url, **params)
 
-    def get(self, wazo_tenant=None, **params):
+    def get(self, token=None, wazo_tenant=None, **params):
         url = str(self)
         params = self._merge_params(params, self.body)
         if wazo_tenant:
             params['headers'] = {'Wazo-Tenant': wazo_tenant}
+        if token:
+            params['headers'] = {'X-Auth-Token': token}
         return self.client.get(url, **params)
 
-    def post(self, body=None, wazo_tenant=None, **params):
+    def post(self, body=None, token=None, wazo_tenant=None, **params):
         url = str(self)
         params = self._merge_params(params, body, self.body)
-        headers = {'Wazo-Tenant': wazo_tenant} if wazo_tenant else None
+        headers = {}
+        if wazo_tenant:
+            headers['Wazo-Tenant'] = wazo_tenant
+        if token:
+            headers['X-Auth-Token'] = token
+        headers = headers or None
         return self.client.post(url, params, headers=headers)
 
-    def put(self, body=None, query_string=None, wazo_tenant=None, **params):
+    def put(self, body=None, query_string=None, token=None, wazo_tenant=None, **params):
         url = str(self)
         params = self._merge_params(params, body, self.body)
-        headers = {'Wazo-Tenant': wazo_tenant} if wazo_tenant else None
+        headers = {}
+        if wazo_tenant:
+            headers['Wazo-Tenant'] = wazo_tenant
+        if token:
+            headers['X-Auth-Token'] = token
+        headers = headers or None
         return self.client.put(url, params, query_string, headers=headers)
 
-    def delete(self, wazo_tenant=None, **params):
+    def delete(self, token=None, wazo_tenant=None, **params):
         url = str(self)
-        headers = {'Wazo-Tenant': wazo_tenant} if wazo_tenant else None
+        headers = {}
+        if wazo_tenant:
+            headers['Wazo-Tenant'] = wazo_tenant
+        if token:
+            headers['X-Auth-Token'] = token
+        headers = headers or None
         return self.client.delete(url, params, headers=headers)
 
     def _copy(self):