+++ wazo_confd/plugins/voicemail_general/resource.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from flask import request
@@ -8,7 +8,7 @@ from marshmallow.validate import Length
 
 from xivo_dao.alchemy.staticvoicemail import StaticVoicemail
 
-from wazo_confd.auth import required_acl
+from wazo_confd.auth import required_acl, required_master_tenant
 from wazo_confd.helpers.mallow import BaseSchema
 from wazo_confd.helpers.restful import ConfdResource
 
@@ -45,11 +45,13 @@ class VoicemailGeneralList(ConfdResource):
         super().__init__()
         self.service = service
 
+    @required_master_tenant()
     @required_acl('confd.asterisk.voicemail.general.get')
     def get(self):
         options = self.service.list()
         return self.schema().dump({'options': options})
 
+    @required_master_tenant()
     @required_acl('confd.asterisk.voicemail.general.update')
     def put(self):
         form = self.schema().load(request.get_json())